+++ sample/src/main/java/io/snabble/testapp/HomeFragment.java
@@ -18,9 +18,11 @@ import androidx.fragment.app.Fragment;
 import org.apache.commons.io.FileUtils;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.List;
 
 import io.snabble.sdk.Project;
+import io.snabble.sdk.Shop;
 import io.snabble.sdk.Snabble;
 import io.snabble.sdk.ui.SnabbleUI;
 
@@ -90,7 +92,7 @@ public class HomeFragment extends Fragment {
         // convenience project switching for debugging - use geofencing or a single project in real apps
         final List<Project> projectList = Snabble.getInstance().getProjects();
         Spinner projects = v.findViewById(R.id.projects);
-        projects.setAdapter(new ArrayAdapter<Project>(requireContext(), R.layout.item_project, projectList) {
+        projects.setAdapter(new ArrayAdapter<Project>(requireContext(), R.layout.item_dropdown, projectList) {
             @NonNull
             @Override
             public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
@@ -121,6 +123,8 @@ public class HomeFragment extends Fragment {
 
                 SnabbleUI.useProject(project);
 
+                updateShops(v);
+
                 if (project.getShops().length > 0) {
                     project.setCheckedInShop(project.getShops()[0]);
                 }
@@ -132,6 +136,52 @@ public class HomeFragment extends Fragment {
             }
         });
 
+        updateShops(v);
+
         return v;
+    }
+
+    private void updateShops(View v) {
+        Project project = SnabbleUI.getProject();
+        final List<Shop> shopList = Arrays.asList(project.getShops());
+        Spinner shops = v.findViewById(R.id.shops);
+        shops.setAdapter(new ArrayAdapter<Shop>(requireContext(), R.layout.item_dropdown, shopList) {
+            @NonNull
+            @Override
+            public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+                TextView v = (TextView) super.getView(position, convertView, parent);
+                Shop shop = shopList.get(position);
+                v.setText(shop.getName() + " (" + shop.getId() + ")");
+                return v;
+            }
+
+            @Override
+            public View getDropDownView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+                TextView v = (TextView) super.getDropDownView(position, convertView, parent);
+                Shop shop = shopList.get(position);
+                v.setText(shop.getName() + " (" + shop.getId() + ")");
+                return v;
+            }
+        });
+
+        for (int i=0; i<shopList.size(); i++) {
+            if (project.getCheckedInShop() == shopList.get(i)) {
+                shops.setSelection(i);
+                break;
+            }
+        }
+
+        shops.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
+                Shop shop = shopList.get(position);
+                project.setCheckedInShop(shop);
+            }
+
+            @Override
+            public void onNothingSelected(AdapterView<?> parent) {
+
+            }
+        });
     }
 }